refactor and add comments
