+++ client/api_bucket.go
@@ -34,10 +34,9 @@ type Bucket interface {
 	HeadBucket(ctx context.Context, bucketName string) (*storageTypes.BucketInfo, error)
 	HeadBucketByID(ctx context.Context, bucketID string) (*storageTypes.BucketInfo, error)
 	// PutBucketPolicy apply bucket policy to the principal, return the txn hash
-	PutBucketPolicy(bucketName string, principalStr types.Principal, statements []*permTypes.Statement, opt types.PutPolicyOption) (string, error)
+	PutBucketPolicy(ctx context.Context, bucketName string, principalStr types.Principal, statements []*permTypes.Statement, opt types.PutPolicyOption) (string, error)
 	// DeleteBucketPolicy delete the bucket policy of the principal
-	DeleteBucketPolicy(bucketName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error)
-	DeleteObjectPolicy(bucketName, objectName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error)
+	DeleteBucketPolicy(ctx context.Context, bucketName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error)
 	// GetBucketPolicy get the bucket policy info of the user specified by principalAddr
 	GetBucketPolicy(ctx context.Context, bucketName string, principalAddr sdk.AccAddress) (*permTypes.Policy, error)
 
@@ -207,7 +206,7 @@ func (c *client) HeadBucketByID(ctx context.Context, bucketID string) (*storageT
 }
 
 // PutBucketPolicy apply bucket policy to the principal, return the txn hash
-func (c *client) PutBucketPolicy(bucketName string, principalStr types.Principal,
+func (c *client) PutBucketPolicy(ctx context.Context, bucketName string, principalStr types.Principal,
 	statements []*permTypes.Statement, opt types.PutPolicyOption) (string, error) {
 	km, err := c.chainClient.GetKeyManager()
 	if err != nil {
@@ -223,11 +222,11 @@ func (c *client) PutBucketPolicy(bucketName string, principalStr types.Principal
 	putPolicyMsg := storageTypes.NewMsgPutPolicy(km.GetAddr(), resource.String(),
 		principal, statements, opt.PolicyExpireTime)
 
-	return c.sendPutPolicyTxn(putPolicyMsg, opt.TxOpts)
+	return c.sendPutPolicyTxn(ctx, putPolicyMsg, opt.TxOpts)
 }
 
 // DeleteBucketPolicy delete the bucket policy of the principal
-func (c *client) DeleteBucketPolicy(bucketName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error) {
+func (c *client) DeleteBucketPolicy(ctx context.Context, bucketName string, principalAddr sdk.AccAddress, opt types.DeletePolicyOption) (string, error) {
 	km, err := c.chainClient.GetKeyManager()
 	if err != nil {
 		return "", errors.New("key manager is nil")
@@ -236,7 +235,7 @@ func (c *client) DeleteBucketPolicy(bucketName string, principalAddr sdk.AccAddr
 	resource := gnfdTypes.NewBucketGRN(bucketName).String()
 	principal := permTypes.NewPrincipalWithAccount(principalAddr)
 
-	return c.sendDelPolicyTxn(km.GetAddr(), resource, principal, opt.TxOpts)
+	return c.sendDelPolicyTxn(ctx, km.GetAddr(), resource, principal, opt.TxOpts)
 }
 
 // IsBucketPermissionAllowed check if the permission of bucket is allowed to the user
